delete other admins for main admin account

+++ delete-admin.php
@@ -4,8 +4,10 @@ include 'db.php';
 ensure_logged_in();
 
 #Query to perform delete
+
 $id=$_GET["id"];
-if ($id != 1){
+$user=$_GET["user"];
+if ($id != 1 && $user != 1){
     mysqli_query($db, "delete from announcements where adminId=$id");
     mysqli_query($db, "delete from Admins where adminId=$id");
 
@@ -14,11 +16,11 @@ if ($id != 1){
     session_destroy();
     header('Location:index.php?accountDeleted');
     exit();
+
 } else {
-    ?>
-    <script type="text/javascript">
-    window.location="edit-admin.php?masterCantBeDeleted";
-    </script>
-    <?php
+    mysqli_query($db, "delete from announcements where adminId=$id");
+    mysqli_query($db, "delete from Admins where adminId=$id");
+    header('Location:edit-admin.php?accountDeletedOther');
+    exit();
 }
     ?>
